ai-description-service/requirements.txt
Flask==3.1.0
uvicorn==0.32.1
pillow==11.0.0
transformers==4.47.0
torch==2.5.1
requests==2.32.3